Move logic from cmd/manager to a package for easier testing.
