+++ cmd/manager/main.go
@@ -24,18 +24,9 @@ import (
 	goruntime "runtime"
 	"time"
 
-	"k8s.io/klog"
-
-	"github.com/spf13/pflag"
-
 	// Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 
-	"k8s.io/apimachinery/pkg/runtime"
-	kubernetesscheme "k8s.io/client-go/kubernetes/scheme"
-	appsv1defaults "k8s.io/kubernetes/pkg/apis/apps/v1"
-	batchv1defaults "k8s.io/kubernetes/pkg/apis/batch/v1"
-	corev1defaults "k8s.io/kubernetes/pkg/apis/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
@@ -43,12 +34,9 @@ import (
 
 	"github.com/operator-framework/operator-sdk/pkg/k8sutil"
 	"github.com/operator-framework/operator-sdk/pkg/leader"
-	"github.com/operator-framework/operator-sdk/pkg/log/zap"
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
 
-	planetscaleapis "planetscale.dev/vitess-operator/pkg/apis"
-	"planetscale.dev/vitess-operator/pkg/controller"
-	vbssubcontroller "planetscale.dev/vitess-operator/pkg/controller/vitessbackupstorage/subcontroller"
+	"planetscale.dev/vitess-operator/pkg/operator/controllermanager"
 	"planetscale.dev/vitess-operator/pkg/operator/fork"
 )
 
@@ -62,30 +50,7 @@ var (
 	metricsHost       = "0.0.0.0"
 	metricsPort int32 = 8383
 )
-var log = logf.Log.WithName("cmd")
-
-// schemeAddFuncs are all the things we register into our compiled-in API type system (Scheme).
-var schemeAddFuncs = []func(*runtime.Scheme) error{
-	// This is the types only (no defaulters) for all built-in APIs supported by client-go.
-	kubernetesscheme.AddToScheme,
-
-	/*
-		These are the defaulters for some built-in APIs that we use.
-
-		This code doesn't come with the Kubernetes client library,
-		so we have to vendor it from the main server repository.
-
-		Doing this lets us achieve `kubectl apply`-like semantics while
-		still using statically-typed Go structs instead of JSON/YAML
-		to define our objects.
-	*/
-	corev1defaults.AddToScheme,
-	appsv1defaults.AddToScheme,
-	batchv1defaults.AddToScheme,
-
-	// This is our own CRDs.
-	planetscaleapis.AddToScheme,
-}
+var log = logf.Log.WithName("manager")
 
 func printVersion() {
 	log.Info(fmt.Sprintf("Go Version: %s", goruntime.Version()))
@@ -98,41 +63,7 @@ func main() {
 	// do something other than run the main operator code.
 	forkPath := fork.Path()
 
-	// Add the zap logger flag set to the CLI. The flag set must
-	// be added before calling pflag.Parse().
-	pflag.CommandLine.AddFlagSet(zap.FlagSet())
-
-	// Add flags registered by imported packages (e.g. glog and
-	// controller-runtime)
-	pflag.CommandLine.AddGoFlagSet(flag.CommandLine)
-
-	pflag.Parse()
-
-	// Initialize flags for klog, which is necessary to configure logging from
-	// the low-level k8s client libraries. We don't use glog ourselves, but we
-	// have dependencies that use it, so we have to follow the instructions for
-	// making klog coexist with glog:
-	// https://github.com/kubernetes/klog/blob/master/examples/coexist_glog/coexist_glog.go
-	klogFlags := flag.NewFlagSet("klog", flag.ExitOnError)
-	klog.InitFlags(klogFlags)
-	// Sync the glog and klog flags.
-	pflag.CommandLine.VisitAll(func(f1 *pflag.Flag) {
-		f2 := klogFlags.Lookup(f1.Name)
-		if f2 != nil {
-			value := f1.Value.String()
-			f2.Value.Set(value)
-		}
-	})
-
-	// Use a zap logr.Logger implementation. If none of the zap
-	// flags are configured (or if the zap flag set is not being
-	// used), this defaults to a production zap logger.
-	//
-	// The logger instantiated here can be changed to any logger
-	// implementing the logr.Logger interface. This logger will
-	// be propagated through the whole operator, generating
-	// uniform and structured logs.
-	logf.SetLogger(zap.Logger())
+	controllermanager.InitFlags()
 
 	printVersion()
 
@@ -161,18 +92,7 @@ func main() {
 		}
 	}
 
-	// Set up scheme for all resources we depend on.
-	scheme := runtime.NewScheme()
-	for _, add := range schemeAddFuncs {
-		if err := add(scheme); err != nil {
-			log.Error(err, "")
-			os.Exit(1)
-		}
-	}
-
-	// Create a new Cmd to provide shared dependencies and start components
-	mgr, err := manager.New(cfg, manager.Options{
-		Scheme:             scheme,
+	mgr, err := controllermanager.New(forkPath, cfg, manager.Options{
 		Namespace:          namespace,
 		SyncPeriod:         cacheInvalidateInterval,
 		MetricsBindAddress: fmt.Sprintf("%s:%d", metricsHost, metricsPort),
@@ -182,32 +102,9 @@ func main() {
 		os.Exit(1)
 	}
 
-	log.Info("Registering Components.")
+	log.Info("Starting the manager.")
 
-	// We use the fork path primarily to decide which controllers to run in this
-	// manager process. Not all controllers run in the root process, for example.
-	switch forkPath {
-	case "":
-		// Run all root controllers, defined as anything that registers itself
-		// in the top-level 'pkg/controller' at package init time.
-		if err := controller.AddToManager(mgr); err != nil {
-			log.Error(err, "")
-			os.Exit(1)
-		}
-	case vbssubcontroller.ForkPath:
-		// Run only the vitessbackupstorage subcontroller.
-		if err := vbssubcontroller.Add(mgr); err != nil {
-			log.Error(err, "")
-			os.Exit(1)
-		}
-	default:
-		log.Error(fmt.Errorf("undefined fork path: %v", forkPath), "")
-		os.Exit(1)
-	}
-
-	log.Info("Starting the Cmd.")
-
-	// Start the Cmd
+	// Start the manager
 	if err := mgr.Start(signals.SetupSignalHandler()); err != nil {
 		log.Error(err, "Manager exited non-zero")
 		os.Exit(1)
